+++ apps/web/src/app/_shared/notifications/api/patch-read-all-notifications.ts
