+++ apps/web/src/app/_shared/notifications/api/use-read-all-notifications.ts
