+++ apps/web/src/app/notifications/_components/NotificationView.tsx
@@ -2,7 +2,7 @@
 
 import { useRouter } from "next/navigation";
 import { ChevronLeft } from "@/shared/ui/icons/ChevronLeft";
-import { useReadAllNotifications } from "../_api/use-read-all-notifications";
+import { useReadAllNotifications } from "../../_shared/notifications/api/use-read-all-notifications";
 import { NotificationBoundary } from "./NotificationBoundary";
 import { NotificationList } from "./NotificationList";
